fixed linking InventoryMenu to InventoryButton without modifying scene
implemented workaround to not modify scene by adding InventoryMenu prefab to InventoryButton prefab. In future if we do a scene clean up I beleive we can get rid of the InventoryMenus at the top level of both scenes, didn't do now because I didn't want to mess with the scene file.